+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 - Fix not being able to group send images when the the image was a link instead of a local path
 - Replace Yandex- with Google-Translate
 - Remove ``/danbooru_latest`` and replace ``/danbooru_search`` with ``/danbooru``
+- Add more danbooru like services. Instead of just danbooru we get safebooru, yandere and konachan too
 
 
 2.4.0 (2019-01-05)